+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeTemplateLabelsEndpoint`
 - Add `EnvelopeEndpoint.discard` and `EnvelopeEndpoint.restore`
 - Add `EnvelopeProperties.signatureTagParts`
+- Add `EnvelopesEndpoint.validate`
 
 ### Changed
 - Replace fields `authenticationMethod`, `authenticationPlace` and `authenticateOnDownload` with `authenticationOnOpen`, `authenticationOnSignature` and `authenticationOnDownload`